+++ tests/test_result/test_result_base.py
@@ -7,6 +7,7 @@ from returns.result import Result
 
 @pytest.mark.parametrize('method_name', [
     'bind',
+    'unify',
     'map',
     'rescue',
     'fix',